+++ src/JP.cpp
@@ -9,6 +9,7 @@
 
 
 #include <Rcpp.h>
+
 using namespace Rcpp;
 
 // [[Rcpp::export]]