+++ src/UnionFind.h
@@ -15,6 +15,7 @@
 #define UNIONFIND
 
 #include <Rcpp.h>
+
 using namespace Rcpp;
 
 class UnionFind